// src/types/ev-charger.ts
import { Operator, Version } from "./index";

export type EVCharger = {
  id: number;
  company: Pick<Operator, "id" | "business_name" | "created_by" | "customer_type">;
  floor_name: unknown;
  buildings: unknown;
  space_name: unknown;
  factories: unknown;
  groups: unknown;
  versions: Version;
  device_type: string;
  device_type_id: number;
  device_type_category_name: string;
  device_type_category_id: number;
  parent_device_id: number | null;
  device_type_icon: string | null;
  name: string | null;
  occupied_ignored: boolean;
  occupancy_timeout: unknown;
  occupied_brightness: number;
  not_occupied_brightness: number;
  oro_id: string;
  sensor_id: unknown;
  led: boolean;
  ble_mac_address: unknown;
  model: unknown;
  serial_number: string;
  version: unknown;
  wifi_mac_address: unknown;
  light_hue: number;
  light_dimming_level: number;
  light_state: boolean;
  occupancy_state: string | null;
  mac_address: unknown;
  major: unknown;
  minor: unknown;
  rssi: unknown;
  uuid: unknown;
  bssid: unknown;
  channel: unknown;
  wifi_rssi: unknown;
  wifi_ssid: unknown;
  wifi_standard: unknown;
  status: boolean;
  last_connection: string | null;
  first_connection: string | null;
  date_mfg: string | null;
  interface: unknown;
  vm_name: string;
  working_status: unknown;
  has_shoppingcart: unknown;
  deleted_at: string | null;
  created_at: string;
  updated_at: string;
  distributor_id: unknown;
  is_virtual: boolean;
  allowShoppingcart: boolean;
  payment_terminal: unknown;
  is_webcheckout: boolean;
  fixtures: unknown;
  location: number;
  themes: string;
  adverts: unknown;
  parent_device: unknown;
  location_details: string;
  longitude: number | null;
  latitude: number | null;
  address_text: string;
  address: {latitude: number, longitude: number, address: string},
  chiller_temperature: number | null;
  percentage_total_stock: number | null;
  device_temperature_data: number | null;
  device_temperature: number | null;
  has_peer_device: boolean;
  peer_device_ids: number[];
  peer_device_details: unknown[];
  max_power: number;
  connector_type: string;
};

export type EmotionT = "age" | "emotion";